requirements.txt
fastapi==0.109.0
PyYAML==6.0.1
uvicorn==0.26.0
